Revert "refactor：SendNotification"
This reverts commit 99a030023c27bfb28916aab032446d597c871a69.
