+++ src/Contracts/Masa.Mc.Contracts.Admin/_Imports.cs
@@ -23,5 +23,5 @@ global using Magicodes.ExporterAndImporter.Core.Models;
 global using System.Collections.ObjectModel;
 global using System.ComponentModel.DataAnnotations;
 global using Masa.Mc.Contracts.Admin.Enums.Subjects;
-global using System.Collections.Concurrent;
 global using Masa.Mc.Infrastructure.Common.Utils;
+global using System.Collections.Concurrent;
