+++ src/Services/Masa.Mc.Service/Application/WebsiteMessages/Commands/SendCheckNotificationCommand.cs
@@ -1,0 +1,8 @@
+﻿// Copyright (c) MASA Stack All rights reserved.
+// Licensed under the Apache License. See LICENSE.txt in the project root for license information.
+
+namespace Masa.Mc.Service.Admin.Application.WebsiteMessages.Commands;
+
+public record SendCheckNotificationCommand : Command
+{
+}
