+++ src/Services/Masa.Mc.Service/Application/WebsiteMessages/Commands/SendGetNotificationCommand.cs
@@ -1,8 +1,8 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the Apache License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.Mc.Service.Admin.Application.Notifications.Commands;
+namespace Masa.Mc.Service.Admin.Application.WebsiteMessages.Commands;
 
-public record SendNotificationCommand(SendNotificationDto dto) : Command
+public record SendGetNotificationCommand(IReadOnlyList<string> UserIds) : Command
 {
 }
